requirements.txt
pytest
requests
allure-pytest
selene>=2.0.b4
python-dotenv